Enabled secure Websocket to allow WebGL to connect to our server
Added a checkbox to allow Secure connection or hosting.
The server can now detect and use the certificate.pfx in the
Data folder as long as the Secure checkbox is on, generated by
openssl pkcs12 -export -out certificate.pfx -inkey private.key -in certificate.crt
It's assumed that Client won't host with Secure, which is also
added to the localhost button functionality, which disables it.
There's currently a bug, such that clients connecting to the
unsecured address while Secured causes them to hang indefinitely.