Fix to prevent duplicate Esc CarSelection, and auto-disable Secure
Pressing Esc key within CarSelection now gets ignored as a part of
addressing a LD Jam issue. We don't want accidental disconnects.

We can't make secure off by default due to our server, yet we also
can't make it on by default, because it blocks custom servers,
making it unintuitive for playera as there's no error message,
so a compromise was chosen to automatically turn off the
Secured when a certificate is missing, or when localhost is chosen.
There's now also the initialization of NetworkSettingsUi for the
situations of player returning from in-game with changed values.
The Disconnect button is now better aligned with the Connect ones.

Also moved the PlayerSpawns positions and marked them in editor.