Detached part cleanup, better idle workflow, Garage & Ping fix
The detached parts/wheels are now properly being registered on
both server and clients, and they are being cleaned up when the
round restarts. However, we still aren't syncing existing parts
to the newly connected players mid-match, due to an open Mirror
issue of being unable to spawn a prefab's children directly.

CarSelection now properly places an AudioListener in the middle of
the Arena, so that players can stay idle and listen for other new
connections without hearing their own engine. On top of that, in
order to further support being idle, the client auto-reconnects
in the event of an accidental disconnect (it only tries once).

Moved Garage away from the Arena due to Additively loaded scene,
so that other players aren't visible, and enabled the vehicle
on Start so that NetworkIdentity doesn't cause it to be removed.

Fixed displaying Ping on client along with the prolonged warnings
by adding it to the NetworkManager's Registered Spawnable Prefabs,
and made it less intrusive by removing the white background.